Move "save_profile" function to middleware
